+++ src/main/java/com/kaltura/client/types/AssetBaseFilter.java
@@ -74,12 +74,12 @@ public abstract class AssetBaseFilter extends RelatedFilter {
 	private String tagsLike;
 	private String tagsMultiLikeOr;
 	private String tagsMultiLikeAnd;
-	private Integer createdAtGreaterThanOrEqual;
-	private Integer createdAtLessThanOrEqual;
-	private Integer updatedAtGreaterThanOrEqual;
-	private Integer updatedAtLessThanOrEqual;
-	private Integer deletedAtGreaterThanOrEqual;
-	private Integer deletedAtLessThanOrEqual;
+	private Long createdAtGreaterThanOrEqual;
+	private Long createdAtLessThanOrEqual;
+	private Long updatedAtGreaterThanOrEqual;
+	private Long updatedAtLessThanOrEqual;
+	private Long deletedAtGreaterThanOrEqual;
+	private Long deletedAtLessThanOrEqual;
 
 	// idEqual:
 	public String getIdEqual(){
@@ -214,10 +214,10 @@ public abstract class AssetBaseFilter extends RelatedFilter {
 	}
 
 	// createdAtGreaterThanOrEqual:
-	public Integer getCreatedAtGreaterThanOrEqual(){
+	public Long getCreatedAtGreaterThanOrEqual(){
 		return this.createdAtGreaterThanOrEqual;
 	}
-	public void setCreatedAtGreaterThanOrEqual(Integer createdAtGreaterThanOrEqual){
+	public void setCreatedAtGreaterThanOrEqual(Long createdAtGreaterThanOrEqual){
 		this.createdAtGreaterThanOrEqual = createdAtGreaterThanOrEqual;
 	}
 
@@ -226,10 +226,10 @@ public abstract class AssetBaseFilter extends RelatedFilter {
 	}
 
 	// createdAtLessThanOrEqual:
-	public Integer getCreatedAtLessThanOrEqual(){
+	public Long getCreatedAtLessThanOrEqual(){
 		return this.createdAtLessThanOrEqual;
 	}
-	public void setCreatedAtLessThanOrEqual(Integer createdAtLessThanOrEqual){
+	public void setCreatedAtLessThanOrEqual(Long createdAtLessThanOrEqual){
 		this.createdAtLessThanOrEqual = createdAtLessThanOrEqual;
 	}
 
@@ -238,10 +238,10 @@ public abstract class AssetBaseFilter extends RelatedFilter {
 	}
 
 	// updatedAtGreaterThanOrEqual:
-	public Integer getUpdatedAtGreaterThanOrEqual(){
+	public Long getUpdatedAtGreaterThanOrEqual(){
 		return this.updatedAtGreaterThanOrEqual;
 	}
-	public void setUpdatedAtGreaterThanOrEqual(Integer updatedAtGreaterThanOrEqual){
+	public void setUpdatedAtGreaterThanOrEqual(Long updatedAtGreaterThanOrEqual){
 		this.updatedAtGreaterThanOrEqual = updatedAtGreaterThanOrEqual;
 	}
 
@@ -250,10 +250,10 @@ public abstract class AssetBaseFilter extends RelatedFilter {
 	}
 
 	// updatedAtLessThanOrEqual:
-	public Integer getUpdatedAtLessThanOrEqual(){
+	public Long getUpdatedAtLessThanOrEqual(){
 		return this.updatedAtLessThanOrEqual;
 	}
-	public void setUpdatedAtLessThanOrEqual(Integer updatedAtLessThanOrEqual){
+	public void setUpdatedAtLessThanOrEqual(Long updatedAtLessThanOrEqual){
 		this.updatedAtLessThanOrEqual = updatedAtLessThanOrEqual;
 	}
 
@@ -262,10 +262,10 @@ public abstract class AssetBaseFilter extends RelatedFilter {
 	}
 
 	// deletedAtGreaterThanOrEqual:
-	public Integer getDeletedAtGreaterThanOrEqual(){
+	public Long getDeletedAtGreaterThanOrEqual(){
 		return this.deletedAtGreaterThanOrEqual;
 	}
-	public void setDeletedAtGreaterThanOrEqual(Integer deletedAtGreaterThanOrEqual){
+	public void setDeletedAtGreaterThanOrEqual(Long deletedAtGreaterThanOrEqual){
 		this.deletedAtGreaterThanOrEqual = deletedAtGreaterThanOrEqual;
 	}
 
@@ -274,10 +274,10 @@ public abstract class AssetBaseFilter extends RelatedFilter {
 	}
 
 	// deletedAtLessThanOrEqual:
-	public Integer getDeletedAtLessThanOrEqual(){
+	public Long getDeletedAtLessThanOrEqual(){
 		return this.deletedAtLessThanOrEqual;
 	}
-	public void setDeletedAtLessThanOrEqual(Integer deletedAtLessThanOrEqual){
+	public void setDeletedAtLessThanOrEqual(Long deletedAtLessThanOrEqual){
 		this.deletedAtLessThanOrEqual = deletedAtLessThanOrEqual;
 	}
 
@@ -307,12 +307,12 @@ public abstract class AssetBaseFilter extends RelatedFilter {
 		tagsLike = GsonParser.parseString(jsonObject.get("tagsLike"));
 		tagsMultiLikeOr = GsonParser.parseString(jsonObject.get("tagsMultiLikeOr"));
 		tagsMultiLikeAnd = GsonParser.parseString(jsonObject.get("tagsMultiLikeAnd"));
-		createdAtGreaterThanOrEqual = GsonParser.parseInt(jsonObject.get("createdAtGreaterThanOrEqual"));
-		createdAtLessThanOrEqual = GsonParser.parseInt(jsonObject.get("createdAtLessThanOrEqual"));
-		updatedAtGreaterThanOrEqual = GsonParser.parseInt(jsonObject.get("updatedAtGreaterThanOrEqual"));
-		updatedAtLessThanOrEqual = GsonParser.parseInt(jsonObject.get("updatedAtLessThanOrEqual"));
-		deletedAtGreaterThanOrEqual = GsonParser.parseInt(jsonObject.get("deletedAtGreaterThanOrEqual"));
-		deletedAtLessThanOrEqual = GsonParser.parseInt(jsonObject.get("deletedAtLessThanOrEqual"));
+		createdAtGreaterThanOrEqual = GsonParser.parseLong(jsonObject.get("createdAtGreaterThanOrEqual"));
+		createdAtLessThanOrEqual = GsonParser.parseLong(jsonObject.get("createdAtLessThanOrEqual"));
+		updatedAtGreaterThanOrEqual = GsonParser.parseLong(jsonObject.get("updatedAtGreaterThanOrEqual"));
+		updatedAtLessThanOrEqual = GsonParser.parseLong(jsonObject.get("updatedAtLessThanOrEqual"));
+		deletedAtGreaterThanOrEqual = GsonParser.parseLong(jsonObject.get("deletedAtGreaterThanOrEqual"));
+		deletedAtLessThanOrEqual = GsonParser.parseLong(jsonObject.get("deletedAtLessThanOrEqual"));
 
 	}
