+++ src/main/java/com/kaltura/client/types/BaseInteractivity.java
@@ -57,7 +57,7 @@ public abstract class BaseInteractivity extends ObjectBase {
 	/**
 	 * Interactivity update date as Unix timestamp (In seconds)
 	 */
-	private Integer updatedAt;
+	private Long updatedAt;
 
 	// data:
 	public String getData(){
@@ -80,7 +80,7 @@ public abstract class BaseInteractivity extends ObjectBase {
 		return this.entryId;
 	}
 	// updatedAt:
-	public Integer getUpdatedAt(){
+	public Long getUpdatedAt(){
 		return this.updatedAt;
 	}
 
@@ -97,7 +97,7 @@ public abstract class BaseInteractivity extends ObjectBase {
 		data = GsonParser.parseString(jsonObject.get("data"));
 		version = GsonParser.parseInt(jsonObject.get("version"));
 		entryId = GsonParser.parseString(jsonObject.get("entryId"));
-		updatedAt = GsonParser.parseInt(jsonObject.get("updatedAt"));
+		updatedAt = GsonParser.parseLong(jsonObject.get("updatedAt"));
 
 	}
