+++ src/main/java/com/kaltura/client/types/BaseSyndicationFeed.java
@@ -100,7 +100,7 @@ public abstract class BaseSyndicationFeed extends ObjectBase {
 	/**
 	 * Creation date as Unix timestamp (In seconds)
 	 */
-	private Integer createdAt;
+	private Long createdAt;
 	/**
 	 * allow_embed tells google OR yahoo weather to allow embedding the video on google
 	  OR yahoo video results   or just to provide a link to the landing page.   it is
@@ -130,7 +130,7 @@ public abstract class BaseSyndicationFeed extends ObjectBase {
 	/**
 	 * Update date as Unix timestamp (In seconds)
 	 */
-	private Integer updatedAt;
+	private Long updatedAt;
 	private Boolean useCategoryEntries;
 	/**
 	 * Feed content-type header value
@@ -202,7 +202,7 @@ public abstract class BaseSyndicationFeed extends ObjectBase {
 	}
 
 	// createdAt:
-	public Integer getCreatedAt(){
+	public Long getCreatedAt(){
 		return this.createdAt;
 	}
 	// allowEmbed:
@@ -326,7 +326,7 @@ public abstract class BaseSyndicationFeed extends ObjectBase {
 	}
 
 	// updatedAt:
-	public Integer getUpdatedAt(){
+	public Long getUpdatedAt(){
 		return this.updatedAt;
 	}
 	// useCategoryEntries:
@@ -372,7 +372,7 @@ public abstract class BaseSyndicationFeed extends ObjectBase {
 		status = SyndicationFeedStatus.get(GsonParser.parseInt(jsonObject.get("status")));
 		type = SyndicationFeedType.get(GsonParser.parseInt(jsonObject.get("type")));
 		landingPage = GsonParser.parseString(jsonObject.get("landingPage"));
-		createdAt = GsonParser.parseInt(jsonObject.get("createdAt"));
+		createdAt = GsonParser.parseLong(jsonObject.get("createdAt"));
 		allowEmbed = GsonParser.parseBoolean(jsonObject.get("allowEmbed"));
 		playerUiconfId = GsonParser.parseInt(jsonObject.get("playerUiconfId"));
 		flavorParamId = GsonParser.parseInt(jsonObject.get("flavorParamId"));
@@ -383,7 +383,7 @@ public abstract class BaseSyndicationFeed extends ObjectBase {
 		entriesOrderBy = SyndicationFeedEntriesOrderBy.get(GsonParser.parseString(jsonObject.get("entriesOrderBy")));
 		enforceEntitlement = GsonParser.parseBoolean(jsonObject.get("enforceEntitlement"));
 		privacyContext = GsonParser.parseString(jsonObject.get("privacyContext"));
-		updatedAt = GsonParser.parseInt(jsonObject.get("updatedAt"));
+		updatedAt = GsonParser.parseLong(jsonObject.get("updatedAt"));
 		useCategoryEntries = GsonParser.parseBoolean(jsonObject.get("useCategoryEntries"));
 		feedContentTypeHeader = GsonParser.parseString(jsonObject.get("feedContentTypeHeader"));
