+++ src/main/java/com/kaltura/client/types/BaseUser.java
@@ -97,11 +97,11 @@ public class BaseUser extends ObjectBase {
 	/**
 	 * Creation date as Unix timestamp (In seconds)
 	 */
-	private Integer createdAt;
+	private Long createdAt;
 	/**
 	 * Last update date as Unix timestamp (In seconds)
 	 */
-	private Integer updatedAt;
+	private Long updatedAt;
 	/**
 	 * Can be used to store various partner related data as a string
 	 */
@@ -112,7 +112,7 @@ public class BaseUser extends ObjectBase {
 	private LanguageCode language;
 	private Integer lastLoginTime;
 	private Integer statusUpdatedAt;
-	private Integer deletedAt;
+	private Long deletedAt;
 	private String allowedPartnerIds;
 	private String allowedPartnerPackages;
 	private UserMode userMode;
@@ -278,11 +278,11 @@ public class BaseUser extends ObjectBase {
 	}
 
 	// createdAt:
-	public Integer getCreatedAt(){
+	public Long getCreatedAt(){
 		return this.createdAt;
 	}
 	// updatedAt:
-	public Integer getUpdatedAt(){
+	public Long getUpdatedAt(){
 		return this.updatedAt;
 	}
 	// partnerData:
@@ -346,7 +346,7 @@ public class BaseUser extends ObjectBase {
 		return this.statusUpdatedAt;
 	}
 	// deletedAt:
-	public Integer getDeletedAt(){
+	public Long getDeletedAt(){
 		return this.deletedAt;
 	}
 	// allowedPartnerIds:
@@ -410,8 +410,8 @@ public class BaseUser extends ObjectBase {
 		tags = GsonParser.parseString(jsonObject.get("tags"));
 		adminTags = GsonParser.parseString(jsonObject.get("adminTags"));
 		status = UserStatus.get(GsonParser.parseInt(jsonObject.get("status")));
-		createdAt = GsonParser.parseInt(jsonObject.get("createdAt"));
-		updatedAt = GsonParser.parseInt(jsonObject.get("updatedAt"));
+		createdAt = GsonParser.parseLong(jsonObject.get("createdAt"));
+		updatedAt = GsonParser.parseLong(jsonObject.get("updatedAt"));
 		partnerData = GsonParser.parseString(jsonObject.get("partnerData"));
 		indexedPartnerDataInt = GsonParser.parseInt(jsonObject.get("indexedPartnerDataInt"));
 		indexedPartnerDataString = GsonParser.parseString(jsonObject.get("indexedPartnerDataString"));
@@ -419,7 +419,7 @@ public class BaseUser extends ObjectBase {
 		language = LanguageCode.get(GsonParser.parseString(jsonObject.get("language")));
 		lastLoginTime = GsonParser.parseInt(jsonObject.get("lastLoginTime"));
 		statusUpdatedAt = GsonParser.parseInt(jsonObject.get("statusUpdatedAt"));
-		deletedAt = GsonParser.parseInt(jsonObject.get("deletedAt"));
+		deletedAt = GsonParser.parseLong(jsonObject.get("deletedAt"));
 		allowedPartnerIds = GsonParser.parseString(jsonObject.get("allowedPartnerIds"));
 		allowedPartnerPackages = GsonParser.parseString(jsonObject.get("allowedPartnerPackages"));
 		userMode = UserMode.get(GsonParser.parseInt(jsonObject.get("userMode")));
