+++ src/main/java/com/kaltura/client/types/BatchJob.java
@@ -93,9 +93,9 @@ public class BatchJob extends ObjectBase {
 
 	private Long id;
 	private Integer partnerId;
-	private Integer createdAt;
-	private Integer updatedAt;
-	private Integer deletedAt;
+	private Long createdAt;
+	private Long updatedAt;
+	private Long deletedAt;
 	private Integer lockExpiration;
 	private Integer executionAttempts;
 	private Integer lockVersion;
@@ -155,15 +155,15 @@ public class BatchJob extends ObjectBase {
 		return this.partnerId;
 	}
 	// createdAt:
-	public Integer getCreatedAt(){
+	public Long getCreatedAt(){
 		return this.createdAt;
 	}
 	// updatedAt:
-	public Integer getUpdatedAt(){
+	public Long getUpdatedAt(){
 		return this.updatedAt;
 	}
 	// deletedAt:
-	public Integer getDeletedAt(){
+	public Long getDeletedAt(){
 		return this.deletedAt;
 	}
 	// lockExpiration:
@@ -535,9 +535,9 @@ public class BatchJob extends ObjectBase {
 		// set members values:
 		id = GsonParser.parseLong(jsonObject.get("id"));
 		partnerId = GsonParser.parseInt(jsonObject.get("partnerId"));
-		createdAt = GsonParser.parseInt(jsonObject.get("createdAt"));
-		updatedAt = GsonParser.parseInt(jsonObject.get("updatedAt"));
-		deletedAt = GsonParser.parseInt(jsonObject.get("deletedAt"));
+		createdAt = GsonParser.parseLong(jsonObject.get("createdAt"));
+		updatedAt = GsonParser.parseLong(jsonObject.get("updatedAt"));
+		deletedAt = GsonParser.parseLong(jsonObject.get("deletedAt"));
 		lockExpiration = GsonParser.parseInt(jsonObject.get("lockExpiration"));
 		executionAttempts = GsonParser.parseInt(jsonObject.get("executionAttempts"));
 		lockVersion = GsonParser.parseInt(jsonObject.get("lockVersion"));
