+++ src/main/java/com/kaltura/client/types/BulkUpload.java
@@ -74,7 +74,7 @@ public class BulkUpload extends ObjectBase {
 	private Long id;
 	private String uploadedBy;
 	private String uploadedByUserId;
-	private Integer uploadedOn;
+	private Long uploadedOn;
 	private Integer numOfEntries;
 	private BatchJobStatus status;
 	private String logFileUrl;
@@ -127,10 +127,10 @@ public class BulkUpload extends ObjectBase {
 	}
 
 	// uploadedOn:
-	public Integer getUploadedOn(){
+	public Long getUploadedOn(){
 		return this.uploadedOn;
 	}
-	public void setUploadedOn(Integer uploadedOn){
+	public void setUploadedOn(Long uploadedOn){
 		this.uploadedOn = uploadedOn;
 	}
 
@@ -316,7 +316,7 @@ public class BulkUpload extends ObjectBase {
 		id = GsonParser.parseLong(jsonObject.get("id"));
 		uploadedBy = GsonParser.parseString(jsonObject.get("uploadedBy"));
 		uploadedByUserId = GsonParser.parseString(jsonObject.get("uploadedByUserId"));
-		uploadedOn = GsonParser.parseInt(jsonObject.get("uploadedOn"));
+		uploadedOn = GsonParser.parseLong(jsonObject.get("uploadedOn"));
 		numOfEntries = GsonParser.parseInt(jsonObject.get("numOfEntries"));
 		status = BatchJobStatus.get(GsonParser.parseInt(jsonObject.get("status")));
 		logFileUrl = GsonParser.parseString(jsonObject.get("logFileUrl"));
